requirements.txt
aiohttp==3.6.2
discord.py==1.5.1
mojang==0.1.2
mystbin.py==0.3.5
DiscordUtils==1.2.4